Refactor: Simplify pg providers by reusing Postgres client across AlloyDB/CloudSQL providers.
