+++ retrieval_service/datastore/providers/alloydb.py
@@ -23,7 +23,7 @@ from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 import models
 
 from .. import datastore
-from .postgres_datastore import PostgresDatastore
+from .postgres import Client as PostgresClient
 
 ALLOYDB_PG_IDENTIFIER = "alloydb-postgres"
 
@@ -41,14 +41,14 @@ class Config(BaseModel, datastore.AbstractConfig):
 
 class Client(datastore.Client[Config]):
     __connector: Optional[AsyncConnector] = None
-    __pg_ds: PostgresDatastore
+    __pg_client: PostgresClient
 
     @datastore.classproperty
     def kind(cls):
         return ALLOYDB_PG_IDENTIFIER
 
     def __init__(self, async_engine: AsyncEngine):
-        self.__pg_ds = PostgresDatastore(async_engine)
+        self.__pg_client = PostgresClient(async_engine)
 
     @classmethod
     async def create(cls, config: Config) -> "Client":
@@ -83,7 +83,7 @@ class Client(datastore.Client[Config]):
         flights: list[models.Flight],
         policies: list[models.Policy],
     ) -> None:
-        await self.__pg_ds.initialize_data(airports, amenities, flights, policies)
+        await self.__pg_client.initialize_data(airports, amenities, flights, policies)
 
     async def export_data(
         self,
@@ -93,17 +93,17 @@ class Client(datastore.Client[Config]):
         list[models.Flight],
         list[models.Policy],
     ]:
-        return await self.__pg_ds.export_data()
+        return await self.__pg_client.export_data()
 
     async def get_airport_by_id(
         self, id: int
     ) -> tuple[Optional[models.Airport], Optional[str]]:
-        return await self.__pg_ds.get_airport_by_id(id)
+        return await self.__pg_client.get_airport_by_id(id)
 
     async def get_airport_by_iata(
         self, iata: str
     ) -> tuple[Optional[models.Airport], Optional[str]]:
-        return await self.__pg_ds.get_airport_by_iata(iata)
+        return await self.__pg_client.get_airport_by_iata(iata)
 
     async def search_airports(
         self,
@@ -111,31 +111,31 @@ class Client(datastore.Client[Config]):
         city: Optional[str] = None,
         name: Optional[str] = None,
     ) -> tuple[list[models.Airport], Optional[str]]:
-        return await self.__pg_ds.search_airports(country, city, name)
+        return await self.__pg_client.search_airports(country, city, name)
 
     async def get_amenity(
         self, id: int
     ) -> tuple[Optional[models.Amenity], Optional[str]]:
-        return await self.__pg_ds.get_amenity(id)
+        return await self.__pg_client.get_amenity(id)
 
     async def amenities_search(
         self, query_embedding: list[float], similarity_threshold: float, top_k: int
     ) -> tuple[list[Any], Optional[str]]:
-        return await self.__pg_ds.amenities_search(
+        return await self.__pg_client.amenities_search(
             query_embedding, similarity_threshold, top_k
         )
 
     async def get_flight(
         self, flight_id: int
     ) -> tuple[Optional[models.Flight], Optional[str]]:
-        return await self.__pg_ds.get_flight(flight_id)
+        return await self.__pg_client.get_flight(flight_id)
 
     async def search_flights_by_number(
         self,
         airline: str,
         number: str,
     ) -> tuple[list[models.Flight], Optional[str]]:
-        return await self.__pg_ds.search_flights_by_number(airline, number)
+        return await self.__pg_client.search_flights_by_number(airline, number)
 
     async def search_flights_by_airports(
         self,
@@ -143,7 +143,7 @@ class Client(datastore.Client[Config]):
         departure_airport: Optional[str] = None,
         arrival_airport: Optional[str] = None,
     ) -> tuple[list[models.Flight], Optional[str]]:
-        return await self.__pg_ds.search_flights_by_airports(
+        return await self.__pg_client.search_flights_by_airports(
             date, departure_airport, arrival_airport
         )
 
@@ -154,7 +154,7 @@ class Client(datastore.Client[Config]):
         departure_airport: str,
         departure_time: str,
     ) -> tuple[Optional[models.Flight], Optional[str]]:
-        return await self.__pg_ds.validate_ticket(
+        return await self.__pg_client.validate_ticket(
             airline, flight_number, departure_airport, departure_time
         )
 
@@ -170,7 +170,7 @@ class Client(datastore.Client[Config]):
         departure_time: str,
         arrival_time: str,
     ):
-        await self.__pg_ds.insert_ticket(
+        await self.__pg_client.insert_ticket(
             user_id,
             user_name,
             user_email,
@@ -186,14 +186,14 @@ class Client(datastore.Client[Config]):
         self,
         user_id: str,
     ) -> tuple[list[Any], Optional[str]]:
-        return await self.__pg_ds.list_tickets(user_id)
+        return await self.__pg_client.list_tickets(user_id)
 
     async def policies_search(
         self, query_embedding: list[float], similarity_threshold: float, top_k: int
     ) -> tuple[list[str], Optional[str]]:
-        return await self.__pg_ds.policies_search(
+        return await self.__pg_client.policies_search(
             query_embedding, similarity_threshold, top_k
         )
 
     async def close(self):
-        await self.__pg_ds.close()
+        await self.__pg_client.close()
